refactor: use existing events architecture
