+++ src/components/FormFields/CustomUrlField.tsx
@@ -5,7 +5,7 @@ import { useContext } from 'react';
 import useBFFContext from '@/components/app/data/hooks/useBFFContext';
 import { FieldContainer } from '@/components/FieldContainer';
 import Field from '@/components/FormFields/Field';
-import { CHECKOUT_STEPS, PLAN_TYPE, TRACKED_FIELDS } from '@/constants/tracking';
+import { CHECKOUT_STEPS, PLAN_TYPE, TRACKED_FIELDS } from '@/constants/events';
 import { useDebounceTracking } from '@/hooks/useDebounceTracking';
 
 import type { UseFormReturn } from 'react-hook-form';
