+++ src/hooks/tests/useFieldTracking.test.ts
@@ -1,7 +1,7 @@
 import { logError } from '@edx/frontend-platform/logging';
 import { renderHook } from '@testing-library/react';
 
-import { TRACKING_EVENT_NAMES } from '@/constants/tracking';
+import EVENT_NAMES from '@/constants/events';
 import { sendEnterpriseCheckoutTrackingEvent } from '@/utils/common';
 
 import { useFieldTracking } from '../useFieldTracking';
@@ -47,7 +47,7 @@ describe('useFieldTracking', () => {
     expect(mockSendEvent).toHaveBeenCalledTimes(1);
     expect(mockSendEvent).toHaveBeenCalledWith({
       checkoutIntentId: 456,
-      eventName: TRACKING_EVENT_NAMES.CHECKOUT_FIELD_BLUR,
+      eventName: EVENT_NAMES.SUBSCRIPTION_CHECKOUT.CHECKOUT_FIELD_BLURRED,
       properties: {
         step: 'plan_details',
         field_name: 'fullName',
@@ -67,7 +67,7 @@ describe('useFieldTracking', () => {
 
     expect(mockSendEvent).toHaveBeenCalledWith({
       checkoutIntentId: null,
-      eventName: TRACKING_EVENT_NAMES.CHECKOUT_FIELD_BLUR,
+      eventName: EVENT_NAMES.SUBSCRIPTION_CHECKOUT.CHECKOUT_FIELD_BLURRED,
       properties: {
         step: 'registration',
         field_name: 'email',
@@ -91,7 +91,7 @@ describe('useFieldTracking', () => {
 
     expect(mockSendEvent).toHaveBeenCalledWith({
       checkoutIntentId: 789,
-      eventName: TRACKING_EVENT_NAMES.CHECKOUT_FIELD_BLUR,
+      eventName: EVENT_NAMES.SUBSCRIPTION_CHECKOUT.CHECKOUT_FIELD_BLURRED,
       properties: {
         step: 'account_details',
         field_name: 'urlSlug',
@@ -152,7 +152,7 @@ describe('useFieldTracking', () => {
 
     expect(mockSendEvent).toHaveBeenCalledWith({
       checkoutIntentId: 200,
-      eventName: TRACKING_EVENT_NAMES.CHECKOUT_FIELD_BLUR,
+      eventName: EVENT_NAMES.SUBSCRIPTION_CHECKOUT.CHECKOUT_FIELD_BLURRED,
       properties: {
         step: 'registration',
         field_name: 'username',
